+++ packages/react-mui-shared/src/components/PangeaFields/FieldsForm/Fields.tsx
@@ -1,5 +1,5 @@
 import { FC } from "react";
-import { Divider, Grid } from "@mui/material";
+import { Divider, Grid2 as Grid } from "@mui/material";
 
 import FieldComponent from "./FieldComponent";
 import { FieldFormSchema, FormFieldLabelProps } from "./types";
@@ -31,7 +31,6 @@ const Fields: FC<{
 
         return (
           <Grid
-            item
             style={{
               // FIXME: This is a hack. We need to update the schema
               //  to be able to specify separate values