+++ packages/react-mui-audit-log-viewer/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2024-11-04
+
+### Fixed
+
+- Remove stray console logs
+
+## [1.0.0] - 2024-11-04
+
+### Added
+
+- Upgrade material-ui packages to version 6
+- "@mui/icons-material": "^6.0.0",
+- "@mui/material": "^6.0.0",
+- "@pangeacyber/react-mui-shared": "1.0.0-beta.0",
+
 ## [0.0.95] - 2024-10-11
 
 ### Fixed